+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - The `float` module gains the `divide` function.
 - The `int` module gains the `divide`, `power`, and `square_root` functions.
-- The `string` module gains the `first`, `last`, `capitalize` and `from` functions.
+- The `string` module gains the `first`, `last`, `capitalize` and `inspect` functions.
 
 ## v0.21.0 - 2022-04-24